+++ src/Resolvers/ExperienceResolver.ts
@@ -1,4 +1,5 @@
 import ExperienceService from "../Services/ExperienceService.js";
+import { requireAdmin } from "../utils/auth.js";
 
 const ExperienceResolver = {
     Query: {
@@ -10,13 +11,16 @@ const ExperienceResolver = {
         }
     },
     Mutation: {
-        createExperience: async (_: unknown, { input }: { input: any }) => {
+        createExperience: async (_: unknown, { input }: { input: any }, { user }: any) => {
+            requireAdmin(user);
             return await ExperienceService.createExperience(input);
         },
-        updateExperience: async (_: unknown, { id, input }: { id: string, input: any }) => {
+        updateExperience: async (_: unknown, { id, input }: { id: string, input: any }, { user }: any) => {
+            requireAdmin(user);
             return await ExperienceService.updateExperience(id, input);
         },
-        deleteExperience: async (_: unknown, { id }: { id: string }) => {
+        deleteExperience: async (_: unknown, { id }: { id: string }, { user }: any) => {
+            requireAdmin(user);
             return await ExperienceService.deleteExperience(id);
         }
     }
